遊戲功能可運作並新增到歷史紀錄

+++ app/src/main/java/tw/edu/pu/nihongo_benkyo/model/database/HistoryDetail.kt
@@ -23,10 +23,10 @@ import androidx.room.*
 )
 data class HistoryDetail(
     @ColumnInfo(name = "history_id")
-    var historyId: Int,
+    var historyId: Long,
 
     @ColumnInfo(name = "question_id")
-    var questionId: Int,
+    var questionId: Long,
 
     @ColumnInfo(name = "answer")
     var answer: String,
